+++ delta_backend/src/utils.py
@@ -1,6 +1,6 @@
 
 from stdnum.verhoeff import validate
-
+   
 def is_valid_simple_snomed(simple_snomed: str) -> bool:
     """
     This utility is designed for reuse and should be packaged as part of a
