+++ delta_backend/tests/test_convert_person_gender.py
@@ -2,7 +2,7 @@ import copy
 import json
 import unittest
 from utils_for_converter_tests import ValuesForTests
-from delta_converter import Converter
+from converter import Converter
 
 
 class TestPersonGenderToFlatJson(unittest.TestCase):
